requirements.txt
pymongo
discord.py
